fix(postgres): BIGSERIAL ids for users/roles + widen User.id to i64

+++ crates/arium-dioxus/src/server.rs
@@ -76,7 +76,7 @@ pub async fn logout() -> Result<()> {
     let actor = auth
         .current_user
         .as_ref()
-        .and_then(|u| if u.anonymous { None } else { Some(u.id as i64) });
+        .and_then(|u| if u.anonymous { None } else { Some(u.id) });
     auth.logout_user();
     if let Some(id) = actor {
         audit
@@ -100,7 +100,7 @@ pub async fn get_current_user_profile() -> Result<UserProfile> {
     let permissions = if user.anonymous {
         Vec::new()
     } else {
-        auth::list_permissions_for_user(&db.0, user.id as i64)
+        auth::list_permissions_for_user(&db.0, user.id)
             .await
             .unwrap_or_default()
     };
@@ -134,11 +134,7 @@ pub async fn get_resource_role(kind: String, id: i64) -> Result<Option<ResourceR
     }
     let role = authority
         .0
-        .role_on(
-            &db.0,
-            user.id as i64,
-            arium::authz::ResourceRef::new(&kind, id),
-        )
+        .role_on(&db.0, user.id, arium::authz::ResourceRef::new(&kind, id))
         .await
         .map_err(|e| ServerFnError::new(e.to_string()))?;
     Ok(role)
@@ -522,7 +518,7 @@ pub async fn begin_mfa_setup() -> Result<MfaSetupView> {
 
     let label = user.email.clone().unwrap_or_else(|| user.username.clone());
 
-    let info = auth::setup_mfa_secret(&db.0, user.id as i64, &label).await?;
+    let info = auth::setup_mfa_secret(&db.0, user.id, &label).await?;
     Ok(MfaSetupView {
         secret_base32: info.secret_base32,
         qr_png_base64: info.qr_png_base64,
@@ -541,7 +537,7 @@ pub async fn confirm_mfa_setup(code: String) -> Result<()> {
     if user.anonymous {
         return Err(ServerFnError::new("Not signed in.").into());
     }
-    let user_id = user.id as i64;
+    let user_id = user.id;
     if auth::enable_mfa(&db.0, user_id, &code).await? {
         audit
             .record(
@@ -569,7 +565,7 @@ pub async fn disable_mfa_for_user() -> Result<()> {
     if user.anonymous {
         return Err(ServerFnError::new("Not signed in.").into());
     }
-    let user_id = user.id as i64;
+    let user_id = user.id;
     auth::disable_mfa(&db.0, user_id).await?;
     audit
         .record(
@@ -595,7 +591,7 @@ pub async fn get_mfa_status() -> Result<MfaStatusView> {
     if user.anonymous {
         return Ok(MfaStatusView::Disabled);
     }
-    Ok(match auth::mfa_status(&db.0, user.id as i64).await? {
+    Ok(match auth::mfa_status(&db.0, user.id).await? {
         auth::MfaStatus::Disabled => MfaStatusView::Disabled,
         auth::MfaStatus::Pending => MfaStatusView::Pending,
         auth::MfaStatus::Enabled => MfaStatusView::Enabled,
@@ -621,7 +617,7 @@ pub async fn create_api_token(name: String) -> Result<CreateApiTokenResponse> {
     if user.anonymous {
         return Err(ServerFnError::new("Not signed in.").into());
     }
-    let user_id = user.id as i64;
+    let user_id = user.id;
 
     let (token, view) = auth::tokens::create_for_user(&db.0, user_id, &name)
         .await
@@ -658,7 +654,7 @@ pub async fn list_api_tokens() -> Result<Vec<ApiTokenView>> {
     if user.anonymous {
         return Err(ServerFnError::new("Not signed in.").into());
     }
-    auth::tokens::list_for_user(&db.0, user.id as i64)
+    auth::tokens::list_for_user(&db.0, user.id)
         .await
         .map_err(|e| ServerFnError::new(e.to_string()).into())
 }
@@ -677,7 +673,7 @@ pub async fn revoke_api_token(token_id: i64) -> Result<()> {
     if user.anonymous {
         return Err(ServerFnError::new("Not signed in.").into());
     }
-    let user_id = user.id as i64;
+    let user_id = user.id;
 
     let revoked = auth::tokens::revoke_for_user(&db.0, user_id, token_id)
         .await
@@ -754,7 +750,7 @@ async fn require_admin_perm(
         .requires(Rights::permission(perm.to_string()))
         .validate(user, &axum::http::Method::GET, Some(db))
         .await
-        .then_some(user.id as i64)
+        .then_some(user.id)
         .ok_or_else(|| ServerFnError::new("You don't have permission for this action.").into())
 }
 
@@ -794,7 +790,7 @@ pub async fn require_resource_dioxus(
     if user.anonymous {
         return Err(ServerFnError::new("Not signed in.").into());
     }
-    let user_id = user.id as i64;
+    let user_id = user.id;
     // The audit-on-denial composition lives in the engine
     // (`require_resource_audited`); this wrapper only contributes the
     // session→user_id resolution above and the dioxus error mapping below.
@@ -1041,7 +1037,7 @@ pub async fn get_account_view() -> Result<AccountView> {
     if user.anonymous {
         return Err(ServerFnError::new("Not signed in.").into());
     }
-    let id = user.id as i64;
+    let id = user.id;
 
     // We still need a couple of bits that aren't on the cached `User`:
     // whether a password is set, whether MFA is enabled, and which OAuth
@@ -1075,7 +1071,7 @@ pub async fn update_display_name(new_name: String) -> Result<()> {
     if user.anonymous {
         return Err(ServerFnError::new("Not signed in.").into());
     }
-    let id = user.id as i64;
+    let id = user.id;
     let trimmed = new_name.trim();
     let value = if trimmed.is_empty() {
         None
@@ -1106,7 +1102,7 @@ pub async fn change_password(current: String, new_password: String) -> Result<()
     if user.anonymous {
         return Err(ServerFnError::new("Not signed in.").into());
     }
-    let id = user.id as i64;
+    let id = user.id;
     let Some(stored) = auth::get_password_hash(&db.0, id).await? else {
         return Err(ServerFnError::new("This account doesn't use a password.").into());
     };
@@ -1136,7 +1132,7 @@ pub async fn delete_my_account() -> Result<()> {
     if user.anonymous {
         return Err(ServerFnError::new("Not signed in.").into());
     }
-    let id = user.id as i64;
+    let id = user.id;
     auth::soft_delete_user(&db.0, id).await?;
     // Record BEFORE logging out so the auth-session still has the user.
     audit
